// frontend/src/api/generated/models/productPerformanceDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */

export interface ProductPerformanceDto {
  sku: string;
  title: string;
  /** @nullable */
  variantName?: string | null;
  units: number;
  revenue: number;
  cogs: number;
  grossProfit: number;
  marginPct: number;
}